% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateTransitService.R
\docType{data}
\name{D4cModels_ls}
\alias{D4cModels_ls}
\title{D4c simulation models}
\format{a list containing 3 components

\describe{
  \item{NormD4_PtQt}{matrix of normalized D4c values by place type and quantile}
  \item{D4SupplyRatio_Ua}{numeric vector of urbanized area average D4c values by urbanized area}
  \item{AveD4cModel_ls}{list containing components for linear model to predict urbanized area average D4c model}
}}
\source{
SimulateTransitService.R script
}
\usage{
D4cModels_ls
}
\description{
Models to predict average D4c value for urbanized area and to the predict the
D4c values for SimBzones from the urbanized area average and from the
SimBzones place types.
}
\keyword{datasets}
