% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateRoadMiles.R
\name{SimulateRoadMiles}
\alias{SimulateRoadMiles}
\title{Calculate freeway lane-miles per capita by Marea.}
\usage{
SimulateRoadMiles(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{SimulateRoadMiles} calculate freeway lane-miles per capita.
}
\details{
This function calculates freeway lane-miles per capita for each Marea.
}
