% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateTransitService.R
\name{SimulateTransitService}
\alias{SimulateTransitService}
\title{Calculate bus equivalent revenue miles per capita by Marea.}
\usage{
SimulateTransitService(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{SimulateTransitService} calculate bus equivalent revenue miles per capita.
}
\details{
This function calculates bus equivalent revenue miles per capita for each
Marea.
}
