% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateTransitService.R
\docType{data}
\name{VehMiFactors_df}
\alias{VehMiFactors_df}
\title{Revenue miles to vehicle miles conversion factors}
\format{A data frame with 8 rows and 2 variables containing factors for
converting revenue miles of various modes to vehicle miles.
Mode names are 2-character codes corresponding to consolidated mode types.
Consolidated mode types represent modes that have similar characteristics and
bus equivalency values. The consolidate mode codes and their meanings are as
follows:
DR = Demand-responsive
VP = Vanpool and similar
MB = Standard motor bus
RB = Bus rapid transit and commuter bus
MG = Monorail/automated guideway
SR = Streetcar/trolley bus/inclined plain
HR = Heavy Rail/Light Rail
CR = Commuter Rail/Hybrid Rail/Cable Car/Aerial Tramway

\describe{
  \item{Mode}{abbreviation for consolidated mode}
  \item{VehMiFactors}{numeric factors for converting revenue miles to
  vehicle miles}
}}
\source{
AssignTransitService.R script.
}
\usage{
VehMiFactors_df
}
\description{
Vehicle mile factors convert revenue miles for various modes to vehicle
miles for those modes.
}
\keyword{datasets}
